xe: ocl: refactor gen9_global_pooling to use off_t

+++ src/gpu/intel/ocl/gen9_global_pooling.cl
@@ -23,10 +23,10 @@
 KERNEL_ATTR
 __kernel void gen9_global_pooling_fwd(
         __global DATA_T *src, __global int *ws, __global DST_DATA_T *dst) {
-    const int mb = get_global_id(0) / C;
-    const int oc = get_global_id(0) % C;
+    const off_t mb = get_global_id(0) / C;
+    const off_t oc = get_global_id(0) % C;
 
-    const uint dst_off = DST_OFF(mb, oc, 0, 0, 0);
+    const off_t dst_off = DST_OFF(mb, oc, 0, 0, 0);
 
 #if ALG_MAX
 #if DT_BF16
@@ -35,7 +35,7 @@ __kernel void gen9_global_pooling_fwd(
     float dst_val = src[0];
 #endif
 #if IS_TRAINING
-    int max_idx = -1;
+    off_t max_idx = -1;
 #endif
 #else
 #if DT_BF16
@@ -45,10 +45,10 @@ __kernel void gen9_global_pooling_fwd(
 #endif
 #endif
 
-    for (int id = 0; id < ID; id++) {
-        for (int ih = 0; ih < IH; ih++) {
-            for (int iw = 0; iw < IW; iw++) {
-                uint src_off = SRC_OFF(mb, oc, id, ih, iw);
+    for (off_t id = 0; id < ID; id++) {
+        for (off_t ih = 0; ih < IH; ih++) {
+            for (off_t iw = 0; iw < IW; iw++) {
+                off_t src_off = SRC_OFF(mb, oc, id, ih, iw);
 #if DT_BF16
                 DEF_ACC_DATA_T val = DATA_TO_REF(src[src_off]);
 #else
@@ -93,33 +93,33 @@ __kernel void gen9_global_pooling_fwd(
 KERNEL_ATTR
 __kernel void gen9_global_pooling_bwd(__global DATA_T *diff_src,
         __global int *ws, __global DATA_T *diff_dst) {
-    const int mb = GWS_GET_MB();
+    const off_t mb = GWS_GET_MB();
 #if IS_VECTORIZED
-    const int c = GWS_GET_C() + get_sub_group_local_id();
+    const off_t c = GWS_GET_C() + get_sub_group_local_id();
 #else
-    const int c = GWS_GET_C();
+    const off_t c = GWS_GET_C();
 #endif
-    const int spatial = GWS_GET_SPATIAL();
+    const off_t spatial = GWS_GET_SPATIAL();
 
     const bool is_in_padded_area = NEED_ZERO_PADDING && (mb >= MB || c >= C);
-    const int dst_off = DST_OFF(mb, c, 0, 0, 0);
+    const off_t dst_off = DST_OFF(mb, c, 0, 0, 0);
 #if ALG_AVG
     // Read dst value only once
     const DATA_T dst_val = diff_dst[dst_off];
 #endif // ALG_AVG
     int ws_val = ws[dst_off];
-    for (int sp_idx = spatial;
+    for (off_t sp_idx = spatial;
             sp_idx < min(spatial + SPATIAL_CHUNK, SPATIAL_DIM); sp_idx++) {
-        const int iw = sp_idx % IW;
-        const int ih = ((sp_idx - iw) % (IH * IW)) / IW;
-        const int id = (sp_idx - iw - ih * IW) / (IH * IW);
+        const off_t iw = sp_idx % IW;
+        const off_t ih = ((sp_idx - iw) % (IH * IW)) / IW;
+        const off_t id = (sp_idx - iw - ih * IW) / (IH * IW);
         DATA_T val_to_write;
         if (is_in_padded_area)
             val_to_write = DATA_ZERO;
         else {
 #if ALG_MAX
             // Read dst value only in case it's going to be used
-            const int current_input_idx = id * IH * IW + ih * IW + iw;
+            const off_t current_input_idx = id * IH * IW + ih * IW + iw;
             if (current_input_idx == ws_val) {
                 val_to_write = diff_dst[dst_off];
             } else {
@@ -130,7 +130,7 @@ __kernel void gen9_global_pooling_bwd(__global DATA_T *diff_src,
             val_to_write = CONVERT_DATA_T(dst_val_f);
 #endif // ALG_MAX
         }
-        const int src_off = SRC_OFF(mb, GWS_GET_C(), id, ih, iw);
+        const off_t src_off = SRC_OFF(mb, GWS_GET_C(), id, ih, iw);
 #if IS_VECTORIZED
         DST_BLOCK_WRITE(&diff_src[src_off], val_to_write);
 #else
